@@ -47,7 +47,7 @@
     "workspace-schematic": "nx workspace-schematic",
     "dep-graph": "nx dep-graph",
     "help": "nx help",
-    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points && npm run symlinks",
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points && npm run symlinks && git update-index --skip-worktree tests/assets/blog-files/page-6.md",
     "cypress:e2e": "cypress run --spec 'tests/cypress/integration/**/*' --browser chrome --headless",
     "cypress:e2e:server": "start-server-and-test scully:run:cypress http://localhost:1668 cypress:e2e",
     "scully:run:cypress": "node ./dist/libs/scully/scully serve --project=sample-blog --tds",
@@ -61,15 +61,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.0.8",
-    "@angular/common": "10.0.8",
-    "@angular/compiler": "10.0.8",
-    "@angular/core": "10.0.8",
-    "@angular/forms": "10.0.8",
-    "@angular/platform-browser": "10.0.8",
-    "@angular/platform-browser-dynamic": "10.0.8",
-    "@angular/platform-server": "10.0.8",
-    "@angular/router": "10.0.8",
+    "@angular/animations": "10.1.3",
+    "@angular/common": "10.1.3",
+    "@angular/compiler": "10.1.3",
+    "@angular/core": "10.1.3",
+    "@angular/forms": "10.1.3",
+    "@angular/platform-browser": "10.1.3",
+    "@angular/platform-browser-dynamic": "10.1.3",
+    "@angular/platform-server": "10.1.3",
+    "@angular/router": "10.1.3",
     "@nrwl/nx-cloud": "^10.0.0",
     "@sentry/node": "^5.20.1",
     "@types/compression": "^1.7.0",
@@ -97,11 +97,11 @@
     "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1000.5",
-    "@angular-devkit/build-ng-packagr": "0.1000.5",
-    "@angular/cli": "10.0.5",
-    "@angular/compiler-cli": "10.0.8",
-    "@angular/language-service": "10.0.8",
+    "@angular-devkit/build-angular": "0.1001.3",
+    "@angular-devkit/build-ng-packagr": "0.1001.3",
+    "@angular/cli": "10.1.3",
+    "@angular/compiler-cli": "10.1.3",
+    "@angular/language-service": "10.1.3",
     "@nrwl/angular": "10.0.10",
     "@nrwl/builders": "^7.8.7",
     "@nrwl/cli": "^10.2.1",
@@ -164,7 +164,7 @@
     "tsickle": "^0.39.1",
     "tslib": "^2.0.1",
     "tslint": "~6.1.3",
-    "typescript": "3.9.7",
+    "typescript": "^4.0.2",
     "wait-on": "^5.2.0",
     "yamljs": "^0.3.0",
     "yargs": "^15.4.1"
