@@ -6,10 +6,11 @@ import { NavListService } from './components/nav-list/nav-list.service';
   selector: 'app-root',
   encapsulation: ViewEncapsulation.None,
   template: `
+    <input type="checkbox" id="mobile-toggle" />
     <nav class="scullyio-nav-header" theme="dark"></nav>
 
     <div class="page-content">
-      <section class="nav-container" *ngIf="showNavlist">
+      <section class="nav-container" [class.landing]="isLandingPage">
         <div class="scullyio-lang-select"></div>
         <!-- <ul class="scullyio-nav-list"></ul> -->
         <nav>
@@ -25,7 +26,7 @@ import { NavListService } from './components/nav-list/nav-list.service';
 })
 export class AppComponent {
   constructor(private router: Router, public nl: NavListService) {}
-  get showNavlist() {
-    return this.router.url !== '/';
+  get isLandingPage() {
+    return this.router.url === '/';
   }
 }
